DEX-761 begin fleshing out the switch
